Homepage
This push is a more finalized version of the homepage.]

I mostly finalized the style and formatting so that the nav bar works both on desktop and mobile. Also made sure that the footer does not behave differently than expected.

The next push will include more content and images on the other pages.