Education and Contact Pages
Got a good base to build on for the education page and finished the contact page. I want to add some more images to the site overall, but I got all of the text/main content in there for this pages. I also started the projects and experience pages and got the outline of what I need to add for later. I also want to add images on those pages later.