Minimal site build - v1.0
I plan to add on to the site, but this is the basis for the entire site with a majority of the information filled in.

I would like to build on the sites design and make it even better both functionally and design wise.

I also plan to add more pictures of different projects and things to actually look at while viewing my site.